fix: resolve high severity code review findings across backend and frontend
- Fix TOCTOU in resetGame: hold both s.mu and gameState.Mu for entire
  reset to prevent client login between client reset and state reset
- Fix sendGameState blocking: use non-blocking send to prevent game loop
  stalling when broadcast channel is full
- Fix phaser range integer division: use float64 arithmetic to prevent
  silent truncation for future ship types with non-divisible values
- Fix explosion distance boundary: change < to <= so players exactly at
  ExplosionDist are hit, matching original Netrek behavior
- Add WebSocket connection limit (128) to prevent memory exhaustion DoS
- Filter broadcast messages with "to" field: only deliver to the
  intended recipient instead of leaking to all clients
- Clamp intercept prediction horizon to 15 ticks to prevent unrealistic
  acceleration extrapolation at long range
- Prevent coordinated cooldown from overriding higher weapon cooldowns,
  which was making bots fire faster than intended
- Clear BotDefenseTarget when threat is gone in non-tournament mode to
  prevent bots from getting permanently stuck in defense state
- Reset game state on WebSocket reconnect to prevent rendering ghost
  entities from previous session

+++ server/bot_weapons.go
@@ -107,7 +107,7 @@ func (s *Server) fireBotPhaser(p *game.Player, target *game.Player) {
 	}
 
 	// Calculate phaser range using original formula: PHASEDIST * phaserdamage / 100
-	myPhaserRange := float64(game.PhaserDist * shipStats.PhaserDamage / 100)
+	myPhaserRange := float64(game.PhaserDist) * float64(shipStats.PhaserDamage) / 100.0
 
 	// Bot aims directly at target
 	course := math.Atan2(target.Y-p.Y, target.X-p.X)
@@ -231,7 +231,7 @@ func (s *Server) fireBotPhaserAtPlasma(p *game.Player, plasma *game.Plasma) bool
 	dist := game.Distance(p.X, p.Y, plasma.X, plasma.Y)
 
 	// Calculate phaser range using original formula
-	myPhaserRange := float64(game.PhaserDist * shipStats.PhaserDamage / 100)
+	myPhaserRange := float64(game.PhaserDist) * float64(shipStats.PhaserDamage) / 100.0
 
 	if dist > myPhaserRange {
 		return false
@@ -306,7 +306,7 @@ func (s *Server) tryPhaserNearbyPlasma(p *game.Player) bool {
 	}
 
 	shipStats := game.ShipData[p.Ship]
-	myPhaserRange := float64(game.PhaserDist * shipStats.PhaserDamage / 100)
+	myPhaserRange := float64(game.PhaserDist) * float64(shipStats.PhaserDamage) / 100.0
 
 	// Check fuel and temperature
 	phaserCost := shipStats.PhaserDamage * shipStats.PhaserFuelMult
@@ -601,7 +601,7 @@ func (s *Server) planetDefenseWeaponLogic(p *game.Player, enemy *game.Player, en
 
 	// Opportunistic phaser usage - prioritize planet protection over fuel conservation
 	// Phasers can be fired in any direction regardless of ship facing
-	myPhaserRange := float64(game.PhaserDist * shipStats.PhaserDamage / 100)
+	myPhaserRange := float64(game.PhaserDist) * float64(shipStats.PhaserDamage) / 100.0
 	if !firedWeapon && enemyDist < myPhaserRange && p.Fuel > 1000 && p.WTemp < shipStats.MaxWpnTemp-100 {
 		// Fire phasers more liberally when defending planets
 		s.fireBotPhaser(p, enemy)
@@ -638,7 +638,7 @@ func (s *Server) starbaseDefenseWeaponLogic(p *game.Player, enemy *game.Player, 
 	// Aggressive phaser usage for planet defense
 	// Phasers can be fired in any direction regardless of ship facing
 	// Use the canonical phaser range formula for consistency
-	sbPhaserRange := float64(game.PhaserDist * shipStats.PhaserDamage / 100)
+	sbPhaserRange := float64(game.PhaserDist) * float64(shipStats.PhaserDamage) / 100.0
 	if enemyDist < sbPhaserRange && p.Fuel > 1500 && p.WTemp < shipStats.MaxWpnTemp-100 {
 		s.fireBotPhaser(p, enemy)
 		p.BotCooldown = 8
